pyproject.toml
[project]
name = "black-tonny-backend-base"
version = "0.1.0"
description = "Clean FastAPI dual-database base repository for future migration work."
authors = [{ name = "Lizhihui", email = "dev@example.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11, <4"
dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic[email]>=2.12.5",
    "fastapi>=0.109.1",
    "python-jose>=3.3.0",
    "arq>=0.25.0",
    "bcrypt>=4.3.0",
    "python-multipart>=0.0.20",
    "redis>=5.0.1",
    "uvicorn>=0.27.0",
    "uvloop>=0.19.0",
    "httptools>=0.7.1",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "SQLAlchemy>=2.0.25",
    "greenlet>=2.0.2",
    "pydantic-settings>=2.12.0",
    "uuid6>=2024.1.12",
    "fastcrud>=0.19.2",
    "gunicorn>=23.0.0",
    "structlog>=25.1.0,<=25.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.2",
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=1.0.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["src/"]

[tool.hatch.build.targets.wheel]
include = ["src/"]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "C",
    "UP",
    "I",
]
ignore = [
    "E402",
    "UP006",
    "UP007",
    "E741",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401",
    "F403",
]

[tool.ruff.lint.mccabe]
max-complexity = 24

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::PendingDeprecationWarning:starlette.formparsers",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = "src.app.*"
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "src.examples.*"
disallow_untyped_defs = true
